resolves install issue in OctoPrint and a couple of other fixes that was causing it to not load.
